requirements.txt
python-telegram-bot==20.6
groq>=0.4.0
python-dotenv>=1.0.0
pydub>=0.25.1
requests>=2.31.0